Issue #1792 Splitted up draft functionality from composer
