+++ extension/js/common/composer/interfaces/composer-ur-params.ts
@@ -1,0 +1,15 @@
+export type ComposerUrlParams = {
+    disableDraftSaving: boolean;
+    isReplyBox: boolean;
+    tabId: string;
+    acctEmail: string;
+    threadId: string;
+    draftId: string;
+    subject: string;
+    from: string | undefined;
+    to: string[];
+    frameId: string;
+    parentTabId: string;
+    skipClickPrompt: boolean;
+    debug: boolean;
+};
